Restructured project
Implemented all current modules as Angular modules and restructured the project to make this more clear
