+++ index.html
@@ -19,13 +19,13 @@
     <script src="libs/angular/angular.min.js"></script>
     <script src="libs/angular-route/angular-route.min.js"></script>
 
-		<!-- APP JS -->
-		<script type="module" src="js/ui.js"></script>
-		<script type="module" src="js/window.js"></script>
-
     <!-- ANGULAR CUSTOM -->
+		<script src="js/controllers/SimulatorCtrl.js"></script>
+		<script src="js/modules/data-utils.js"></script>
+		<script src="js/modules/ui.js"></script>
+		<script src="js/modules/window-utils.js"></script>
+		<script src="js/modules/window.js"></script>
     <script src="js/controllers/MainCtrl.js"></script>
-    <script src="js/controllers/SimulatorCtrl.js"></script>
     <script src="js/controllers/HomebrewCtrl.js"></script>
     <script src="js/appRoutes.js"></script>
     <script src="js/app.js"></script>
