+++ public/js/modules/interface.js
@@ -1,0 +1,4 @@
+var jsInterface = angular.module('jsInterface', [])
+
+jsInterface.ADMIN = 1
+var ADMIN = 1
