+++ public/views/simulator.html
